miRegionCopy(): handle realloc failure better
Zero out the region size when freeing the region so callers don't think
there's anything there.   (Pointer is already set to NULL from the realloc
result itself.)  Return 0 to the callers, and have them cascade that back
to their callers to indicate failure, instead of their usual return value
of 1 on success.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Backported-to-NX-by: Ulrich Sibiller <uli42@gmx.de>