miRegionOp(): ensure region size is not updated if realloc fails
This function performs operations on a region, and when finished,
checks to see if it should compact the rectangle list.  If the
number of rectangles for which memory is allocated in the list is
more than twice the number used, it tries to shrink.   realloc()
should not fail in this case, but if it does, might as well keep
the correct value for the number of allocated rectangles, so we
don't try to grow it unnecessarily later if adding to the region.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Backported-to-NX-by: Ulrich Sibiller <uli42@gmx.de>