PresentRegion->PresentPixmap, define options, add UST mode
Changes the name of the PresentPixmap request from PresentRegion in
anticipation of future additions of non-pixmap sourced updates (YUV
images in particular).

Adds definitions for all of the new PresentPixmap options.

Adds PresentQueryCapabilities to provide applications the ability to
learn what the underlying hardware can support. One requirement for
any capability is that the X server must do something sensible even if
the client behaves as if a capability is supported when it is not.

Adds IdleNotify events. As pixmaps can go idle in any order, it's
important for applications to know which pixmap to use next. We cannot
use fences as the fence itself may not be signaled for some time after
the X server has figured out which pixmap to idle.

Note that the encoding and header files are not entirely up to date now.

Signed-off-by: Keith Packard <keithp@keithp.com>