nx-X11: handle source pictures (those without a Drawable surface) gracefully.
Cherry-picked from branch 3.5.0.x.

This is basically a merge of the most current xorg-server (1.17.1) code
into nx-X11.

It makes sure that for source pictures, which do not have a drawable
surface, a filter is selected that is supported on the "main" and all
other screens.  Alternatively, if the requested filter is not available
on all screens and the picture is a source picture, this function fails
gracefully.

Additionally, the ChangePictureFilter hook is now called for non-source
pictures.

This also needs an implementation in mipict.{c,h}. The default hook does
nothing and returns a success value.