libnx-xinerama1: also create libXinerama symlink in libnx-xinerama1.postinst.postinst (and remove in libnx-xinerama1.postinst.prerm.)
Due to the nx-x11-common package being a noarch/allarch package,
creating the symlink in nx-libs' Makefile will lead to the symlink
referencing the "default" architecture dpkg uses for building
noarch/allarch packages.

Incidentally, this worked fine for Debian, as amd64 seems to be the
default architecture. On Ubuntu, however, the default architecture up to
Vivid (15.04) was i386. For those builds, the symlink pointed to the 32
bit library of libNX_Xinerama.so.1 -- essentially breaking this feature.

Move the symlink creation to the arch-sensitive libnx-xinerama1 package.
The postinst and prerm scriptlets will work fine, unless someone
installs the i386 package version *after* the amd64 version. Given that
we already create symlinks to libNX_X11 and friends using that method,
no new regression is introduced. Strictly speaking that's a bug, but
we'll hopefully clean that up later...