This patch allows to cleanup the nxcomp resources to allow for a second connection inside the same process, instead of a new process as is the nxproxy case.
This involves creating a new API call

void NXTransCleanupForReconnect(void);

which basically cleans up the global state for the connection but does not exit the process.

Background
==========
This is needed for the IOS platform, where the nxproxy model of forking does not work.
Also NX handles most of the errors with an "exit" call which in IOS cannot be easily handled.