Xinerama: do not cut off at outer edges
This fixes the problem Mike Gabriel describes like this:

- Launch a session on a system with a single monitor. Enable Xinerama
  for this session.
- Open a desktop session in that session window (e.g. MATE or XFCE).
- Move the NX/MATE-or-XFCE session window around on that one
  monitor. Bump at the borders, so that the session window moves into
  the invisible parts around your monitor.

What you see is that the MATE-or-XFCE window manager will become
really busy with resizing the windows and panels in the NX session,
because moving the window over the physical borders of the display
will trigger resize events.

This is non-intuitive, I think.

Same with multi-monitors on the outside edges of the physical Xorg
RandR setup.