Remove compatibility code for nxcomp before 3.5.0
As a part of nxcomp's clean-up process, compatibility with old protocol
versions has been set now at 3.5.0.

This commit removes compatibility code from nxcomp so the minimum supported
version changes from 1.5.0 (protocol step 6) to 3.5.0 (protocol step 10).

At this moment the removal is not complete and it will be followed by some
other commits, in order to get rid of several components and files that are
no longer used as they were only needed to work with old protocol versions.

Refs: ArcticaProject/nx-libs#108