nx-X11/programs/Xserver/hw/nxagent/Handlers.h: fix stack smashing related to different data type sizes between Xserver and Xlib.
While on the X server side we were expecting a 32bit value, the Xlib side in
nxcompext wrote a 64bit integer to the 32bit location, hence overwriting
"random" data on the stack (most notably the return adress.)

We can work around this by using the Xlib-based Window data type on the
server-side as well, but this problem is likely a more general one.

Calling functions from libcompext in nxagent sounds like a bad idea. Other code
locations might well be affected by issues alike to that as well.

In order to silence compiler warnings, use a "macro hack" for overriding the
"Window" type in NXlib.h.