Keyboard.c: evdev is only existing on Linux
Remove check for the remote side operating system because the evdev
check will only succeed on Linux (and in future FreeBSD?) anyway.
Besides, this makes keycode conversion work when using nxagent without
nx transport (as an xnest replacement) or when not providing
client=linux in the options.

This fixes the second issue mentioned in ArcticaProject/nx-libs#239
(keyboard not working correctly).