Revert "Switch from using libNX_X11's deprecated XKeycodeToKeysym() function to using XGetKeyboardMapping()."
This reverts commit efc0dae0519aa0ef1fabea6a64919475fd916347.

Recent test revealed keyboard hangs on high latency connections. These
hangs are not happening in 3.5.0 releases if the nx-libs.

The commit above is responsible for them so we take that back.

Interestingly nxcomp has special treatment for XGetKeyboardMapping()
so it should normally speed up things but it results in the opposite
for us. Needs further examination.

This fixes ArcticaProject/nx-libs#450