Remove maximum size limits for nxagent window
Set maximum size to MAXSHORT (32767).

This especially helps when reconnecting to a session that had
initially been started from a client with a small(er) screen. Before
it was impossible to increase the maximum size of a running session
beyond the screen size of the client that had initiated the session.

This change, together with the previous commit, fixes ArcticaProject/nx-libs#472