nxagent: bypass X2go's keyboard configuration hack
If X2go runs in auto keyboard mode it will pass keyboard=null/null to
the agent and set the keyboard afterwards with setxkbmap. This patch
lets nxagent handle that situation internally (null/null will be
interpreted as clone) and disables X2go's mechanism (by creating a dir
that effectively blocks it; see x2gosetkeyboard)

This is only activated if the agent is run as "x2goagent".

Fixes ArcticaProject/nx-libs#368