treat options parameter as option string if it starts with nx/nx
This way one can easily pass options to nxagent run as Xnest
replacement without having to create a temporary options file. Please
note that it is not of much use for normal sessions. Options passed
like this cannot be changed after startup so a reconnect may not work
as expected.

Fixes ArcticaProject/nx-libs#476