Bump to Mesa 7.0.4+
Mesa has no tag for the 7.0.4 release, so we use the latest commit in
the mesa_7_0_branch (which is 77a8e26bf824b45c1244f5699d7cbfca0fe0a48a).

Unfortunately we are going _down_ version wise:

Building nx-libs with Mesa 7.0 results in glxinfo reporting:
  OpenGL version string: 1.2 (2.1 Mesa 7.0)
But building with Mesa 7.0.4 results in
  OpenGL version string: 1.2 (2.0 Mesa 7.0.4)

Reason: 7.0.4 disables one OpenGL extension essential for OpenGL 2.1
called GL_ARB_shading_language_120 (because it is not feature
complete).