nxcomp, nxproxy: don't use symlinks
Compilation under Windows/cygwin was broken due to the incomplete
symlink support on that platform.

nxcomp/VERSION and nxproxy/VERSION were both symbolic links pointing
to VERSION on their parent directory, but cygwin git was just writting
to regular files instead of symlinks but with the symlink path as their
contents (instead of the target file contents!)

This patch just avoids the symlinks making the configure script look
for the version information on real VERSION file on the parent
directory.