Allow options to be enclosed in '' "" or ||
Note: Nesting and escaping is not supported currently! Quoting is is
only allowed around the complete value part of an option. Quoting must
start directly after the '=' and end directly before ',' or EOS.

This is not perfect, e.g. the following string will be processed erroneously:

key="v,a,l",u,e1",key2=value2

However, it will help if some value needs to contain a comma.