add option keyconv=(auto|on|off)
Adds a new option called "keyconv" to control keycode conversion.

Before commit 2f2ade61a8823bad012737b2b388dcc168cccbbf keycode
conversion was activated if the client was Linux and client side rules
and/or model was "evdev". The only (and undocumented) way to disable
that was providing a value different from "linux" for the "client"
option (which had no other effect). The mentioned commit removed the
dependency on Linux and so there was no way anymore to disable keycode
conversion.