Args.c: do not handle -options twice
An options string/filename can only be provided via the command line
or environment. Both are handled in the one-time init part of
ddxProcessArguments() and will not change during runtime. As we do not
want an options file to include another "options" option there's no
need to handle that a second time. Due to the (kind of crazy) way
nxagent handles all this option stuff we cannot remove the second
check completely.

This also removes the 1024 byte limit for nxagentOptionsFilenameOrString.