Revert "NXhsm.c: add missing code"
This reverts commit c804d99f2c4ea21e37cc8f6c3168b7b79d0d721a.

It seems like the missing code was missing for a good reason. Including
it leads to very strange issues when coupled with libXcomp and using
connection speed settings lower than LAN (and even on LAN some icons
are not showing up correctly, e.g., when using MATE).

Further investigation on why this happens pending and might happen at a
later time.

Fixes: ArcticaProject/nx-libs#656