Events.c: catch intermediate window position changes
nxagentHandleConfigureNotify() has an optimization that accumulates
ConfigureNotify events from the queue to only perform the changes of
the last ConfigureNotify event in the queue. But that code used to
ignore position changes and only adapt the new window position if the
last event happened to contain a position change.

This change ensures the latest position change - if any - found in the
queue will be applied after the accumulation.

Fixes: ArticaProject/nx-libs#688 (second part)