Logger.h: remove pragma, correct attributes
Instead of using a pragma which will be valid for the whole file this
only affects the functions where it is actually required and supresses
"format string is not a string literal" warnings.

According to GCC documentation the second attribute parameter should
be 0:

"For functions where the arguments are not available to be checked
(such as vprintf), specify the third parameter as zero".