nxdialog/: create (almost) empty m4 directory.
Older autoconf versions - as used on SLE 11 and RHEL 6 - choke if
aclocal is passed an include directory that does not exist.

Instead of removing the option from the aclocal call, let's keep it and
just use an empty directory for now. Maybe we'll actually use it at a
later time.