NXmiexpose.c: use upstream version of miPaintWindows()
miPaintWindow() was identical to the version in miexpose.c except
for some unitialized variable fixes. As these also should be in
upstream code we add them there (Note: Xorg never fixed this but
totally rewrote the miPaintWindow() later on.)

This allows us to totally drop our special version of miPaintWindow().