damage: Don't rewrite Text ops to GlyphBlt ops
This backport of a 2013 Xorg fix changes miext/damage.c so that the
text functions (damageText, damage{Poly,Image}Text{8,16}) finally are
very similar to the changes that NX introduced in 2007 with
nxagent-3.0.0-28.

Fun fact: the original freedesktop bug mentioned below is from 2005
and the fix came in 2013...

Next commit will be the complete drop of NXdamage.c.

  commit 0f3f8e4295d4c05f33b93e0ffd2f6f8cd5759bf7
  Author: Adam Jackson <ajax@redhat.com>
  Date:   Wed Aug 21 11:13:11 2013 -0400

    damage: Don't rewrite Text ops to GlyphBlt ops

    There's no particularly good reason to, and it breaks Xnest.

    Bugzilla: http://bugs.freedesktop.org/2454
    Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>
    Reviewed-by: Eric Anholt <eric@anholt.net>
    Signed-off-by: Adam Jackson <ajax@redhat.com>