Clipboard.c: improve XChangeProperty code
The previously used "Atom" type was confusing. Clarify this by
writing the code as requested in the manpage of XChangeProperty:

"If the specified format is 8, the property data must be a char array.
If the specified format is 16, the property data must be a short
array.  If the specified format is 32, the property data must be a
long array."