Window.c: suppress warning
This used to be printed only in TEST mode. Some while ago I had
changed that to WARNING (because it is a warning...). However, this
happens e.g. when running the xscreensaver vfeedback module and it
does not look like it is a problem at all. So let's suppress this
warning again and leave it to the TEST mode as it used to be.