Fix Xfixes event handling
Calling the callback on receptions of an EXTERNAL xfixes event was a weird idea from the
start. However, since adding the protection trap around it it made no sense at all because
it effectivly made the callback a noop for that case.

So let's drop all the trap nonsense and implement it properly. The callback will only be
used for actions by internal clients.